requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mysql-connector-python==8.2.0
networkx==3.2.1
numpy==1.26.2
Pillow==10.1.0
protobuf==4.21.12
requests==2.31.0
sympy==1.12
# torch==2.1.1+cpu
# torchvision==0.16.1+cpu
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
